Create drawImage01.html
Added file: Based on a script I found at W3Schools which loads in a single image by using a button click; I further changed and adapted that script to suit myself...this script uses a thumbnail image click to load in multiple images, one at a time. NOTE: There is still a bug...as each image will over-lap on top of the other one. Need to find a way to clear the Canvas to start again with a 'blank' new screen.